move precalculation function to separate file
